requirements.txt
connexion[flask,swagger-ui]==2.14.2
uvicorn==0.23.2
Flask-SocketIO==5.2.0
python-decouple==3.8
pandas==2.1.1
websockets==12.0
